+++ packages/abap-deploy-config-writer/src/types.ts
@@ -1,0 +1,4 @@
+export interface DeployConfigOptions {
+    baseFile?: string; // e.g ui5.yaml
+    deployFile?: string; // e.g ui5-deploy.yaml
+}