+++ packages/project-access/test/project/dependencies.test.ts
@@ -1,7 +1,9 @@
 import { join, sep } from 'path';
 import type { Package } from '../../src';
-import { getNodeModulesPath } from '../../src';
-import { hasDependency } from '../../src/project/dependencies';
+import { FileName, getNodeModulesPath } from '../../src';
+import { addPackageDevDependency, hasDependency } from '../../src/project/dependencies';
+import { create as createStorage } from 'mem-fs';
+import { create } from 'mem-fs-editor';
 
 describe('Test hasDependency()', () => {
     test('Test package.json has dependency', async () => {
@@ -38,5 +40,29 @@ describe('Test getNodeModulesPath()', () => {
 
     test('Find node_modules parent for relative path, should return undefined', () => {
         expect(getNodeModulesPath(join('some/relative/path'))).toBe(undefined);
+    });
+});
+
+describe('Test updatePackageScript()', () => {
+    const sampleRoot = join(__dirname, '../test-data/json/package');
+
+    test('should add package script', async () => {
+        const fs = create(createStorage());
+        await addPackageDevDependency(sampleRoot, '@mock-lib', '0.0.1', fs);
+        expect(fs.dump(join(sampleRoot, FileName.Package))).toMatchInlineSnapshot(`
+            Object {
+              "": Object {
+                "contents": "{
+                \\"name\\": \\"test\\",
+                \\"version\\": \\"1.0.0\\",
+                \\"devDependencies\\": {
+                    \\"@mock-lib\\": \\"0.0.1\\"
+                }
+            }
+            ",
+                "state": "modified",
+              },
+            }
+        `);
     });
 });