+++ packages/cards-editor-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cards-editor-config-writer
 
+## 0.3.8
+
+### Patch Changes
+
+-   Updated dependencies [399d2ad8]
+    -   @sap-ux/ui5-config@0.22.9
+
 ## 0.3.7
 
 ### Patch Changes